requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
sqlalchemy==2.0.40
python-jose[cryptography]==3.4.1
passlib[argon2]==1.7.4
python-multipart==0.0.20
pydantic==2.10.6
python-dotenv==1.0.1
openai==2.26.0
chromadb==0.6.3
pypdf==5.3.0
